pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fragfm"
version = "0.1.0"
description = "FragFM: Hierarchical Framework for Efficient Molecule Generation via Fragment-Level Discrete Flow Matching"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "scipy==1.14.1",
  "numpy==1.26.4",
  "pandas==2.2.3",
  "scikit-learn==1.5.2",
  "torch==2.1.0",
  "torch-geometric",
  "rdkit==2023.9.2",
  "descriptastorus @ git+https://github.com/bp-kelley/descriptastorus.git",
  "wandb==0.18.6",
  "lmdb==1.5.1",
  "PyYAML==6.0.1",
  "easydict==1.13",
  "parmap==1.7.0",
  "matplotlib==3.8.1",
  "seaborn==0.13.0",
]

[project.optional-dependencies]
dev = ["pre-commit"]

[tool.setuptools.packages.find]
where = ["."]
include = ["fragfm*"]
